third_party/WebKit/LayoutTests/fast/dom/Document/invalid-domain-change-throws-exception.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("This test checks that a SECURITY_ERR exception is raised if an attempt is made to change document.domain to an invalid value.");

shouldThrow('document.domain = "apple.com"', '"SecurityError: Failed to set the \'domain\' property on \'Document\': \'apple.com\' is not a suffix of \'\'."');
</script>
</body>
</html>
