third_party/blink/web_tests/platform/generic/fast/dom/Document/invalid-domain-change-throws-exception-expected.txt

This test checks that a SECURITY_ERR exception is raised if an attempt is made to change document.domain to an invalid value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.domain = "apple.com" threw exception SecurityError: Failed to set the 'domain' property on 'Document': 'apple.com' is not a suffix of ''..
PASS successfullyParsed is true

TEST COMPLETE
